Update rubocop's target ruby version (#18)
Ruby was bumped to 3.1 (jruby 9.4.7.0) in a [recent
PR](https://github.com/elastic/crawler/pull/16).

This PR is a follow-up. It was just the result of calling `rubocop
--autocorrect` on the codebase after updating rubocop's target ruby
version.

The biggest lint change that impacts the codebase between 2.6 and 3.1 is
the introduction of [hash literal
values](https://github.com/rubocop/ruby-style-guide?tab=readme-ov-file#hash-literal-values).