Add retry to bulk indexer (#39)
Related to https://github.com/elastic/crawler/issues/37

We currently have no retry logic for bulk indexing. This is a problem
when running the Crawler on Mac because of an intermittent
`bad_record_mac` error. Retrying often resolves the error (so this is a
bandaid, not a fix).
Retries are healthy in general so this was a good opportunity to add
them.

- Add retry with max of 3 for bulk indexer
- Add timeout
- Add stats for docs that failed indexing
- Improve logging in general