Pin gems and set some platforms as jruby (#91)
In an effort to fix the renovate PRs, this PR tries to set the platform
for specific gems to `:jruby`. This will stop the renovate bot from
fetching these gems from the MRI ruby source.

I decided which gems to add the platform to by looking at all gems with
the `-java` prefix in
https://github.com/elastic/crawler/blob/main/Gemfile.lock

I also saw a good opportunity to alphabetize the gems and pin a version
for every dependency (excluding test or dev dependencies).