Add binary content extraction (#74)
Add binary content extraction support.
When a binary content file (such as a PDF) is encountered, Crawler will
download the file and create an ES doc for ingestion.
The binary content of the file is added as a base64-encoded string to
the `_attachment` field. The actual decoding is done through ES
pipelines, which takes the decoded value and adds it to the `body` field
on the doc. The `_attachment` field is then removed.