Write Event and system logs to log files (#238)
### Closes https://github.com/elastic/search-team/issues/7275

#### This work is part of the larger effort to get Open Crawler to
General Availability (GA)

#### This PR will achieve the following:
- Create a `CrawlLogger` class that not only abstracts logging
implementation away from the `Config` class, but provides an interface
for the creation of new log handlers in the future.
- Add the following user-configurable settings:
1. `log_file_rotation_policy`. Default is 'weekly', accepts 'daily' |
'weekly' | 'monthly'
2. `log_file_directory`. The directory in which to save log files.
Default is ".", accepts any valid directory path
3. `system_logs_to_file`. Whether to write system logs to file. Default
is `false`.
4. `event_logs_to_file`. Whether to write event logs to file. Default is
`false`.
- Amend `EventGenerator` to write events without calling a function in
`Config`
- Amend some existing spec files to work with these changes
- Add new spec files to test the new CrawlLogging code
- Add a basic `filebeat.yml.example` under the `configs/` directory to
allow users to quickly start ingesting logs to ES

The intention with this PR is to implement these changes in a
transparent way - that is, there should be very minimal to no outward
impact to how Crawler functions, apart from when a user opts-in to
having logs get written to file

### Checklists

<!--You can remove unrelated items from checklists below and/or add new
items that may help during the review.-->

#### Pre-Review Checklist
- [x] This PR does NOT contain credentials of any kind, such as API keys
or username/passwords (double check `crawler.yml.example` and
`elasticsearch.yml.example`)
- [x] This PR has a meaningful title
- [x] This PR links to all relevant GitHub issues that it fixes or
partially addresses
- If there is no GitHub issue, please create it. Each PR should have a
link to an issue
- [x] this PR has a thorough description
- [x] Covered the changes with automated tests
- [x] Tested the changes locally
- [x] Added a label for each target release version (example: `v0.1.0`)
- [ ] Considered corresponding documentation changes
- [x] Contributed any configuration settings changes to the
configuration reference