New CLI command to test single URLs with a given config (#262)
### Closes https://github.com/elastic/crawler/issues/182 and
https://github.com/elastic/search-team/issues/9709

This PR aims to add a new CLI command that allows users to test a single
endpoint with a given Crawler URL. This command will output details
regarding whether or not the crawl was successful, an explanation of
failures, and the result of the crawl.

Requisite documentation will be offered as as separate PR once this is
close to completion

### Example command and thought process:

Format:
bin/crawler **urltest** _crawler.yml_ **URL** --es_config
_elasticsearch.yml_

Example command and output
```
bin/crawler urltest config/my-crawler.yml https://www.speedhunters.com/2025/01/project-964-hitting-the-touge-for-the-first-time-in-rwb-form/

[2025-04-10T09:26:10.806Z] [crawl:67f7c6f2714375360db0a1b8] [primary] Initialized an in-memory URL queue for up to 10000 URLs
[2025-04-10T09:26:10.810Z] [crawl:67f7c6f2714375360db0a1b8] [primary] ES connections will be authorized with configured API key
[2025-04-10T09:26:10.811Z] [crawl:67f7c6f2714375360db0a1b8] [primary] ES connections will use SSL without ca_fingerprint
[2025-04-10T09:26:11.645Z] [crawl:67f7c6f2714375360db0a1b8] [primary] Connected to ES at https://ddc74a4e6dfb4f278f7208b1b907d151.us-west2.gcp.elastic-cloud.com:443 - version: 9.0.0-rc1; build flavor: default
[2025-04-10T09:26:11.925Z] [crawl:67f7c6f2714375360db0a1b8] [primary] Index [opencrawler_learning_docker_ci] was found!
[2025-04-10T09:26:11.925Z] [crawl:67f7c6f2714375360db0a1b8] [primary] Elasticsearch sink initialized for index [opencrawler_learning_docker_ci] with pipeline [search-default-ingestion]
[2025-04-10T09:26:12.009Z] [crawl:67f7c6f2714375360db0a1b8] [primary] Crawl status: queue_size=0, pages_visited=0, urls_allowed=0, urls_denied={}, crawl_duration_msec=33, crawling_time_msec=0.0, avg_response_time_msec=0, active_threads=1, http_client={:max_connections=>100, :used_connections=>0}, status_codes={}
[2025-04-10T09:26:15.099Z] [crawl:67f7c6f2714375360db0a1b8] [primary] Crawl queue is empty, finishing the primary crawl
[2025-04-10T09:26:15.099Z] [crawl:67f7c6f2714375360db0a1b8] [primary] Finished a crawl stage. Result: success; Successfully finished the primary crawl with an empty crawl queue
[2025-04-10T09:26:15.100Z] [crawl:67f7c6f2714375360db0a1b8] [primary] Finished a crawl. Result: failure; Successfully finished the primary crawl with an empty crawl queue |

---- URL Test Results ----
- Attempted to crawl https://www.speedhunters.com/2025/01/project-964-hitting-the-touge-for-the-first-time-in-rwb-form/
- Status code: 200
- Content type: text/html; charset=UTF-8
- Crawl duration (seconds): 2.8990111351013184
- Extracted links:
  - http://store.speedhunters.com
  - http://store.speedhunters.com
  - http://www.speedhunters.com/category/carfeatures/
  - http://www.speedhunters.com/tag/car-spotlight/
  - http://www.speedhunters.com/tag/dragracing/
  - https://www.speedhunters.com
  - https://www.speedhunters.com
  - https://www.speedhunters.com/2025/01/project-964-hitting-the-touge-for-the-first-time-in-rwb-form/#content
  - https://www.speedhunters.com/category/content/
  - https://www.speedhunters.com/category/content/special-feature/

You can find the downloaded document under ./crawled_docs
```
The idea here is that we want the provided URL to run through as close
to the 'real crawl' as we can get, short of:
- Following links extracted from the tested URL
- Indexing the resultant docs to ES (if ES is indeed configured)

As such, I've attempted to provide an alternative 'entry point' into the
Crawling process that will heed all given `crawler.yaml` and
`elasticsearch.yml` config options _except_ for the `output_sink` - this
is forced to `File` with a default, top-directory-level folder that
saves the downloaded document for further review. (EDIT - I should add,
this default file location gets overridden if it's present in the
`crawler.yml` - the output will appropriately tell the user the correct
location)

Ignoring ES configs is acceptable because indexing into ES is not the
point of this command - ensuring a given URL is crawl-able with the
given config files is.

I've very much tried to stay away from re-inventing the wheel w.r.t. the
Crawling process, but I can understand if this creates some additional
overhead in how the code 'reads', and would love to hear any feedback on
this front, too.

### Checklists

<!--You can remove unrelated items from checklists below and/or add new
items that may help during the review.-->

#### Pre-Review Checklist
- [x] This PR does NOT contain credentials of any kind, such as API keys
or username/passwords (double check `crawler.yml.example` and
`elasticsearch.yml.example`)
- [x] This PR has a meaningful title
- [x] This PR links to all relevant GitHub issues that it fixes or
partially addresses
- If there is no GitHub issue, please create it. Each PR should have a
link to an issue
- [x] this PR has a thorough description
- [x] Covered the changes with automated tests
- [x] Tested the changes locally
- [x] Added a label for each target release version (example: `v0.1.0`)
- [ ] Considered corresponding documentation changes
- [ ] Contributed any configuration settings changes to the
configuration reference