Fix urltest erroneously marking successful crawls as failures due to lack of purge stage (#300)
### Closes https://github.com/elastic/crawler/issues/299

This PR aims to fix `urltest` erroneously marking successful crawls as
failures by modifying the `Crawler::API::Crawl.combined_outcome()`
function to ignore `results[:purge][:outcome]` and simply return the
result of `results[:primate][:outcome]`.

In other words, if we are in a `urltest` command, we should ignore the
results of a purge crawl because no purge crawl was executed.

### Checklists
#### Pre-Review Checklist
- [x] This PR does NOT contain credentials of any kind, such as API keys
or username/passwords (double check `crawler.yml.example` and
`elasticsearch.yml.example`)
- [x] This PR has a meaningful title
- [x] This PR links to all relevant GitHub issues that it fixes or
partially addresses
- If there is no GitHub issue, please create it. Each PR should have a
link to an issue
- [x] this PR has a thorough description
- [x] Covered the changes with automated tests
- [x] Added a label for each target release version (example: `v0.1.0`)