File sink generates human-readable filenames (#311)
### Closes https://github.com/elastic/crawler/issues/298

This PR aims to make the File sink generate human-readable `json` files
as opposed to the hash values it was previously using. The hash values
were not human readable, which makes understanding their contents at a
glance impossible.

Here is an example of the filenames with this change:

![1033E9FF-73E0-422C-AA31-0610C0FD3A12](https://github.com/user-attachments/assets/75807074-bfac-4211-9890-0f6c4cdab641)

### Checklists

<!--You can remove unrelated items from checklists below and/or add new
items that may help during the review.-->

#### Pre-Review Checklist
- [x] This PR does NOT contain credentials of any kind, such as API keys
or username/passwords (double check `crawler.yml.example` and
`elasticsearch.yml.example`)
- [x] This PR has a meaningful title
- [x] This PR links to all relevant GitHub issues that it fixes or
partially addresses
- If there is no GitHub issue, please create it. Each PR should have a
link to an issue
- [x] this PR has a thorough description
- [x] Covered the changes with automated tests
- [x] Tested the changes locally
- [x] Added a label for each target release version (example: `v0.1.0`)
- [x] Considered corresponding documentation changes